refs 124934: convert ExternalAccount model to ts
