+++ apps/app/src/interfaces/external-account.ts
@@ -1,11 +1,12 @@
-import { Ref } from '@growi/core';
+import { HasObjectId, Ref } from '@growi/core';
 
 import { IUser } from '~/interfaces/user';
 
 
-export type IExternalAccount<ID = string> = {
-  _id: ID,
+export type IExternalAccount = {
   providerType: string,
   accountId: string,
   user: Ref<IUser>,
 }
+
+export type IExternalAccountHasId = IExternalAccount & HasObjectId
